requirements.txt
about-time==4.2.1
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alive-progress==3.1.5
annotated-types==0.7.0
asttokens==2.4.1
async-timeout==5.0.1
attrs==23.2.0
autocfg==0.0.8
autograd==1.6.2
av==13.1.0
bitblas==0.1.0.post1
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.1
cma==3.2.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cpplint==2.0.2
cupy @ file:///home/conda/feedstock_root/build_artifacts/cupy-split_1729277572208/work
cycler==0.12.1
Cython==3.0.11
cython_bbox==0.1.3
datasets==4.0.0
deap==1.4.2
debugpy==1.8.5
decorator==5.1.1
deepspeed==0.17.2
Deprecated==1.2.14
dill==0.3.8
docker-pycreds==0.4.0
docutils==0.21.2
dtlib==0.0.0.dev2
easydict==1.13
einops==0.8.1
exceptiongroup==1.2.2
execnet==2.1.1
executing==2.0.1
fastrlock @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_fastrlock_1734873426/work
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.7.0
fsspec==2024.6.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
gluoncv==0.10.5.post0
grapheme==0.6.0
grpcio==1.66.0
hf-xet==1.1.5
hjson==3.1.0
huggingface-hub==0.33.4
idna==3.7
imageio==2.36.0
imageio-ffmpeg==0.6.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jstyleson==0.0.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.3
neuron_cuda==0.0.0
ninja==1.11.1.1
nncf==2.11.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
openvino-telemetry==2024.1.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
portalocker==2.10.1
progress==1.6
prompt_toolkit==3.0.47
propcache==0.3.2
protobuf==5.27.2
psutil==6.0.0
ptyprocess==0.7.0
PuLP==3.0.2
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyaml==25.1.0
pyarrow==21.0.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydot==3.0.1
Pygments==2.18.0
pymoo==0.6.1.1
pyparsing==3.1.2
pytest==8.3.3
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.1.0
randaugment==1.0.2
RapidFuzz==3.13.0
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.1
scikit-optimize==0.10.2
scipy==1.14.0
seaborn==0.13.2
sentry-sdk==2.19.2
setproctitle==1.3.4
shapely==2.0.6
simclr==1.0.2
six==1.16.0
smmap==5.0.1
spikingjelly==0.0.0.0.14
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==3.0.1
thefuzz==0.22.1
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.9.20
tilelang==0.1.5
timm==0.5.4
tomli==2.0.1
torch==2.4.0
torchaudio==2.4.0
torchinfo==1.8.0
torchstat==0.0.7
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wandb==0.19.0
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.16.0
xxhash==3.5.0
yacs==0.1.8
yarl==1.20.1
